+++ src/pages/SideEssays/SideEssayDetailPage.jsx
@@ -2,8 +2,8 @@ import { useState, useEffect, useContext } from "react";
 import { TitleContext } from "../../components/TitleBar/TitleContext";
 import { useParams, Link } from "react-router-dom";
 import { useLoggedInUser } from "../../components/LoggedInUserContext/LoggedInUserContext";
-import AddCommentForm from '../../components/CommentSection/AddCommentForm';
-import CommentDisplaySection from '../../components/CommentSection/CommentDisplaySection';
+import AddCommentForm from '../../components/Discussion/AddCommentForm';
+import CommentDisplaySection from '../../components/Discussion/CommentDisplaySection';
 import FeedbackMessage from '../../components/FeedbackMessage/FeedbackMessage';
 import { getCommentsOn } from "../../utilities/comments-service";
 import { getSideEssay, starEssayById, unstarEssayById } from "../../utilities/essays-service";
